+++ Runtime/Core/PropertyDrawers/ReadOnlyAttribute.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace Vertx
+{
+	public class ReadOnlyAttribute : PropertyAttribute { }
+}